+++ TemperatureDataBaseFirmware/src/main.cpp
@@ -6,6 +6,7 @@
 
 //---- Scope of functions -----
 float GetTempCelsiusLM35(int adcValue);
+void SendTempSerial(unsigned long delay_ms);
 
 //---- Initial Settings ----
 void setup() 
@@ -16,9 +17,7 @@ void setup()
 //---- Main loop ----
 void loop() 
 {
-    int adc = analogRead(PIN_TEMPERATURE);
-    float temperature = GetTempCelsiusLM35(adc);  
-    Serial.println(temperature);                   //debug
+    SendTempSerial(1000);    //Sends temperature every second
 }
 
 //---- Development of functions ----
@@ -26,4 +25,18 @@ float GetTempCelsiusLM35(int adcValue)
 {
     float voltage = (adcValue * 5.0) / 1023.0;  //Voltage = (ADC * maximum input voltage) / maximum ADC value
     return voltage / 0.010;                     //The lm35 sensor is linear: 10mV/°C
+}
+void SendTempSerial(unsigned long delay_ms)
+{
+    static unsigned long timeLastRead = 0;
+
+    if((millis() - timeLastRead) >= delay_ms)
+    {
+        int adcValue = analogRead(PIN_TEMPERATURE);
+        float temperature = GetTempCelsiusLM35(adcValue);
+        
+        Serial.print("T:"); 
+        Serial.println(temperature);    
+        timeLastRead = millis();
+    }    
 }